Opt into Ubuntu language packs (using dh-sequence-gnome)
Ubuntu's gnome-pkg-tools also runs and depends on dh-translations.
dh-translations doesn't exist in Debian so this is a workaround
to ensure that Ubuntu has the build dependency for dh-translations
and runs it.

Additionally, there is a second workaround supplied by
Sebastien Bacher since dh-translations otherwise wasn't
working correctly for this package